# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: BookCon API
  version: "1.0"
  description: Self-hosted eBook library & sync API. Auth via Bearer JWT (15 min) + rotating refresh tokens. Deletions are tombstones; every syncable entity carries updated_at/deleted_at.
servers:
  - url: "{base}/api/v1"
    variables:
      base:
        default: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    since:
      name: since
      in: query
      schema: { type: string, format: date-time }
      description: Only entities with updated_at > since.
...
  schemas:
...
    Locator:
      description: Readium Locator object (https://readium.org/architecture/models/locators/)
      type: object
      required: [href, type]
      properties:
        href: { type: string }
        type: { type: string }
        title: { type: string }
        locations: { type: object, additionalProperties: true }
        text: { type: object, additionalProperties: true }
    Annotation:
      type: object
      required: [book_id, type, locator]
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        type: { type: string, enum: [highlight, underline, area] }
        locator: { $ref: "#/components/schemas/Locator" }
        color: { type: string, default: yellow }
        note: { type: string, default: "" }
        annotation_tags: { type: array, items: { type: string }, default: [] }
        excerpt: { type: string, default: "" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Bookmark:
      type: object
      required: [book_id, locator]
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        locator: { $ref: "#/components/schemas/Locator" }
        label: { type: string, default: "" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Position:
      type: object
      required: [book_id, locator]
      properties:
        book_id: { type: string, format: uuid }
        locator: { $ref: "#/components/schemas/Locator" }
        progress_percent: { type: number, nullable: true }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /annotations:
    get:
      parameters:
        - { name: book_id, in: query, schema: { type: string, format: uuid } }
        - $ref: "#/components/parameters/since"
      responses:
        "200": { description: List, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Annotation" } } } } }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Annotation" }
      responses:
        "201": { description: Created }
  /annotations/{annotation_id}:
    patch:
      parameters: [{ name: annotation_id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                color: { type: string }
                note: { type: string }
                annotation_tags: { type: array, items: { type: string } }
      responses:
        "200": { description: Updated }
    delete:
      parameters: [{ name: annotation_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Tombstoned }
  /bookmarks:
    get:
      parameters:
        - { name: book_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200": { description: List }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Bookmark" }
      responses:
        "201": { description: Created }
  /bookmarks/{bookmark_id}:
    patch:
      parameters: [{ name: bookmark_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Updated }
    delete:
      parameters: [{ name: bookmark_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Tombstoned }
  /positions/{book_id}:
    put:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Position" }
      responses:
        "200": { description: Stored (LWW), content: { application/json: { schema: { $ref: "#/components/schemas/Position" } } } }
...